feat: Introduce types package
This commit introduces the `types` package, which provides useful types that implement the `sql.Scanner` and `driver.Valuer` interfaces. These types are suitable for use as scan and value targets with the `database/sql` package.

The package includes the following types:

- `BitBool`: An implementation of a bool for the MySQL type BIT(1).
- `JSONText`: A json.RawMessage that validates the JSON format in the source.
- `NullJSONText`: Represents a JSONText that may be null.
- `GzippedText`: A []byte which transparently gzips data being submitted to a database and ungzips data being scanned from a database.

Each type implements the `Value` and `Scan` methods to satisfy the `driver.Valuer` and `sql.Scanner` interfaces, respectively. The package also includes comprehensive tests to ensure the correctness and reliability of its features.