Add go.mod and go.sum files
This commit adds the `go.mod` and `go.sum` files to manage dependencies for the sqlx package.

The `go.mod` file specifies the module path and lists the required dependencies, including:

- github.com/go-sql-driver/mysql
- github.com/i9si-sistemas/assert
- github.com/lib/pq
- github.com/mattn/go-sqlite3

The `go.sum` file contains the cryptographic checksums of the dependencies to ensure the integrity and reproducibility of the build.